Add allowedAttempts setting #1

+++ packages/better-auth/src/plugins/magic-link/index.ts
@@ -16,6 +16,12 @@ interface MagicLinkopts {
 	 */
 	expiresIn?: number;
 	/**
+	 * Allowed attempts for verifying the magic link token.
+	 * Note: Passing Infinity will allow unlimited attempts.
+	 * @default 1
+	 */
+	allowedAttempts?: number;
+	/**
 	 * Send magic link implementation.
 	 */
 	sendMagicLink: (
@@ -64,6 +70,7 @@ interface MagicLinkopts {
 export const magicLink = (options: MagicLinkopts) => {
 	const opts = {
 		storeToken: "plain",
+		allowedAttempts: 1,
 		...options,
 	} satisfies MagicLinkopts;
 
@@ -182,7 +189,7 @@ export const magicLink = (options: MagicLinkopts) => {
 					await ctx.context.internalAdapter.createVerificationValue(
 						{
 							identifier: storedToken,
-							value: JSON.stringify({ email, name: ctx.body.name }),
+							value: JSON.stringify({ email, name: ctx.body.name, attempt: 0 }),
 							expiresAt: new Date(
 								Date.now() + (opts.expiresIn || 60 * 5) * 1000,
 							),
@@ -351,13 +358,33 @@ export const magicLink = (options: MagicLinkopts) => {
 						);
 						throw ctx.redirect(`${errorCallbackURL}?error=EXPIRED_TOKEN`);
 					}
-					await ctx.context.internalAdapter.deleteVerificationValue(
-						tokenValue.id,
-					);
+					const {
+						attempt = 0,
+					} = JSON.parse(tokenValue.value) as {
+						email: string;
+						name?: string | undefined;
+						attempt?: number | undefined;
+					};
+					if (attempt >= opts.allowedAttempts) {
+						await ctx.context.internalAdapter.deleteVerificationValue(
+							tokenValue.id,
+						);
+						throw ctx.redirect(`${errorCallbackURL}?error=ATTEMPTS_EXCEEDED`);
+					}
 					const { email, name } = JSON.parse(tokenValue.value) as {
 						email: string;
 						name?: string;
 					};
+					await ctx.context.internalAdapter.updateVerificationValue(
+						tokenValue.id,
+						{
+							value: JSON.stringify({
+								email,
+								name,
+								attempt: attempt + 1,
+							}),
+						},
+					);
 					let isNewUser = false;
 					let user = await ctx.context.internalAdapter
 						.findUserByEmail(email)
